requirements.txt
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
anthropic>=0.40,<1.0
httpx>=0.27,<1.0
